+++ include/snapclient_dsp_store.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "snapclient_dsp.h"
-
-app_config::SnapclientDspConfig loadSnapclientDspConfig();
-void saveSnapclientDspConfig(const app_config::SnapclientDspConfig &config);